Switch to JSON config and add support for changing update ownership
Android has an unfortunate limitation where the update ownership can
only be set on the initial install. This commit adds support for
changing the owner after the fact. This makes it possible, for example,
to allow an app to only be updatable by an F-Droid client instead of the
Play Store.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>